requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
openpyxl>=3.1.2
pytesseract>=0.3.10
Pillow>=10.1.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
python-multipart>=0.0.6
asyncpg>=0.29.0
alembic>=1.13.0
pandas>=2.1.0
numpy>=1.24.0
python-dotenv>=1.0.0
openai>=2.0.0
tiktoken>=0.5.0
chardet>=5.0.0
colorlog>=6.8.0
pyjwt>=2.8.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
rapidfuzz>=3.0.0
charset-normalizer>=3.0.0
pdf2image>=1.16.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
